autocomplet filter text ok , date fix error when clean
