+++ Front/app/ns_modules/ns_grid/custom.text.autocomplete.filter2.js
@@ -1,0 +1,289 @@
+/**
+ * ag-grid - Advanced Data Grid / Data Table supporting Javascript / React / AngularJS / Web Components
+ * @version v5.4.0
+ * @link http://www.ag-grid.com/
+ * @license MIT
+ */
+
+ define([
+ ], function() {
+
+   'use strict';
+
+    //var utils_1 = require('../utils');
+
+    var utils_1 = {
+        Utils: {}
+    };
+
+    utils_1.Utils.loadTemplate = function (template) {
+        var tempDiv = document.createElement("div");
+        tempDiv.innerHTML = template;
+        return tempDiv.firstChild;
+    };
+
+    utils_1.Utils.addChangeListener = function (element, listener) {
+        element.addEventListener("changed", listener);
+        element.addEventListener("paste", listener);
+        element.addEventListener("input", listener);
+        // IE doesn't fire changed for special keys (eg delete, backspace), so need to
+        // listen for this further ones
+        element.addEventListener("keydown", listener);
+        element.addEventListener("keyup", listener);
+    };
+
+    utils_1.Utils.removeElement = function (parent, cssSelector) {
+        this.removeFromParent(parent.querySelector(cssSelector));
+    };
+    utils_1.Utils.removeFromParent = function (node) {
+        if (node && node.parentNode) {
+            node.parentNode.removeChild(node);
+        }
+    };
+
+    utils_1.Utils.makeNull = function (value) {
+            if (value === null || value === undefined || value === "") {
+                return null;
+            }
+            else {
+                return value;
+            }
+        };
+
+var template = '<div>' +
+    '<div>' +
+    '<select class="ag-filter-select" id="filterType">' +
+    '<option value="1">[CONTAINS]</option>' +
+    '<option value="2">[EQUALS]</option>' +
+    '<option value="3">[NOT EQUALS]</option>' +
+    '<option value="4">[STARTS WITH]</option>' +
+    '<option value="5">[ENDS WITH]</option>' +
+    '<option value="7">In</option>' +
+    '</select>' +
+    '</div>' +
+    '<div>' +
+    '<input class="ag-filter-filter js-text-autocomp" id="filterText" type="text" placeholder="[FILTER...]"/>' +
+    '</div>' +
+    '<div class="ag-filter-apply-panel" id="applyPanel">' +
+    '<button type="button" id="applyButton">[APPLY FILTER]</button>' +
+    '</div>' +
+    '</div>';
+var CONTAINS = 1;
+var EQUALS = 2;
+var NOT_EQUALS = 3;
+var STARTS_WITH = 4;
+var ENDS_WITH = 5;
+var IN = 7;
+return (function () {
+    function TextFilterAutoComp() {
+    }
+    TextFilterAutoComp.prototype.init = function (params) {
+        console.log(params);
+        if (typeof(params.filterParams) === 'undefined' || typeof(params.filterParams.tabToOrder) === 'undefined') {
+            console.error("Text filter autocom need filterParams : { tabToOder : ['string'] } ");
+            return null;
+        }
+        this.filterParams = params.filterParams;
+        this.applyActive = this.filterParams && this.filterParams.apply === true;
+        this.filterChangedCallback = params.filterChangedCallback;
+        this.filterModifiedCallback = params.filterModifiedCallback;
+        this.localeTextFunc = params.localeTextFunc;
+        this.valueGetter = params.valueGetter;
+        console.log(params);
+        this.tabOrdered = params.filterParams.tabToOrder.sort();
+        console.log(this.tabToOrdered);
+        this.createGui();
+        this.filterText = null;
+        this.filterType = CONTAINS;
+        this.createApi();
+
+    };
+    TextFilterAutoComp.prototype.onNewRowsLoaded = function () {
+        var keepSelection = this.filterParams && this.filterParams.newRowsAction === 'keep';
+        if (!keepSelection) {
+            this.api.setType(CONTAINS);
+            this.api.setFilter(null);
+        }
+    };
+    TextFilterAutoComp.prototype.afterGuiAttached = function () {
+        this.eFilterTextField.focus();
+    };
+    TextFilterAutoComp.prototype.doesFilterPass = function (node) {
+        if (!this.filterText) {
+            return true;
+        }
+        var value = this.valueGetter(node);
+        if (!value) {
+            return false;
+        }
+
+
+        var valueLowerCase = value.toString().toLowerCase();
+        switch (this.filterType) {
+            case CONTAINS:
+                return valueLowerCase.indexOf(this.filterText) >= 0;
+            case EQUALS:
+                return valueLowerCase === this.filterText;
+            case NOT_EQUALS:
+                return valueLowerCase != this.filterText;
+            case STARTS_WITH:
+                return valueLowerCase.indexOf(this.filterText) === 0;
+            case ENDS_WITH:
+                var index = valueLowerCase.lastIndexOf(this.filterText);
+                return index >= 0 && index === (valueLowerCase.length - this.filterText.length);
+            case IN:
+                var tab = this.filterText.split(',');
+                if(tab.length <= 1){
+                    tab = this.filterText.split(';');
+                }
+                if(tab.length <= 1){
+                    tab = this.filterText.split(' ');
+                }
+
+                for (var i=0; i< tab.length; i++){
+                    if ((tab[i].toLowerCase() == valueLowerCase)) {
+                        return true;
+                    }
+                }
+                return false;
+            default:
+                // should never happen
+                console.warn('invalid filter type ' + this.filterType);
+                return false;
+        }
+    };
+    TextFilterAutoComp.prototype.getGui = function () {
+        return this.eGui;
+    };
+    TextFilterAutoComp.prototype.isFilterActive = function () {
+        return this.filterText !== null;
+    };
+    TextFilterAutoComp.prototype.createTemplate = function () {
+        return template
+            .replace('[FILTER...]', this.localeTextFunc('filterOoo', 'Filter...'))
+            .replace('[EQUALS]', this.localeTextFunc('equals', 'Equals'))
+            .replace('[NOT EQUALS]', this.localeTextFunc('notEquals', 'Not equals'))
+            .replace('[CONTAINS]', this.localeTextFunc('contains', 'Contains'))
+            .replace('[STARTS WITH]', this.localeTextFunc('startsWith', 'Starts with'))
+            .replace('[ENDS WITH]', this.localeTextFunc('endsWith', 'Ends with'))
+            .replace('[APPLY FILTER]', this.localeTextFunc('applyFilter', 'Apply Filter'));
+    };
+    TextFilterAutoComp.prototype.createGui = function () {
+        var _this = this;
+
+        this.eGui = utils_1.Utils.loadTemplate(this.createTemplate());
+        this.eFilterTextField = this.eGui.querySelector("#filterText");
+        this.eTypeSelect = this.eGui.querySelector("#filterType");
+        utils_1.Utils.addChangeListener(this.eFilterTextField, this.onFilterChanged.bind(this));
+        this.eTypeSelect.addEventListener("change", this.onTypeChanged.bind(this));
+        this.$autoComp = this.eGui.querySelector('.js-text-autocomp');
+        $( this.$autoComp ).autocomplete({
+          source: function(req, responseFn) {
+              var re = $.ui.autocomplete.escapeRegex(req.term);
+              var matcher = new RegExp( "^" + re, "i" );
+              var a = $.grep( _this.tabOrdered, function(item,index){
+                  return matcher.test(item);
+              });
+              responseFn( a );
+          }
+          /*source: _this.tabOrdered*/
+        });
+        this.setupApply();
+    };
+    TextFilterAutoComp.prototype.setupApply = function () {
+        var _this = this;
+        if (this.applyActive) {
+            this.eApplyButton = this.eGui.querySelector('#applyButton');
+            this.eApplyButton.addEventListener('click', function () {
+                _this.filterChangedCallback();
+            });
+        }
+        else {
+            utils_1.Utils.removeElement(this.eGui, '#applyPanel');
+        }
+    };
+    TextFilterAutoComp.prototype.onTypeChanged = function () {
+        this.filterType = parseInt(this.eTypeSelect.value);
+        this.filterChanged();
+    };
+    TextFilterAutoComp.prototype.onFilterChanged = function () {
+        var filterText = utils_1.Utils.makeNull(this.eFilterTextField.value);
+        if (filterText && filterText.trim() === '') {
+            filterText = null;
+        }
+        var newFilterText;
+        if (filterText !== null && filterText !== undefined) {
+            newFilterText = filterText.toLowerCase();
+        }
+        else {
+            newFilterText = null;
+        }
+        if (this.filterText !== newFilterText) {
+            this.filterText = newFilterText;
+            this.filterChanged();
+        }
+    };
+    TextFilterAutoComp.prototype.filterChanged = function () {
+        this.filterModifiedCallback();
+        if (!this.applyActive) {
+            this.filterChangedCallback();
+        }
+    };
+    TextFilterAutoComp.prototype.createApi = function () {
+        var that = this;
+        this.api = {
+            EQUALS: EQUALS,
+            NOT_EQUALS: NOT_EQUALS,
+            CONTAINS: CONTAINS,
+            STARTS_WITH: STARTS_WITH,
+            ENDS_WITH: ENDS_WITH,
+            setType: function (type) {
+                that.filterType = type;
+                that.eTypeSelect.value = type;
+            },
+            setFilter: function (filter) {
+                filter = utils_1.Utils.makeNull(filter);
+                if (filter) {
+                    that.filterText = filter.toLowerCase();
+                    that.eFilterTextField.value = filter;
+                }
+                else {
+                    that.filterText = null;
+                    that.eFilterTextField.value = null;
+                }
+            },
+            getType: function () {
+                return that.filterType;
+            },
+            getFilter: function () {
+                return that.filterText;
+            },
+            getModel: function () {
+                if (that.isFilterActive()) {
+                    return {
+                        type: that.filterType,
+                        filter: that.filterText
+                    };
+                }
+                else {
+                    return null;
+                }
+            },
+            setModel: function (dataModel) {
+                if (dataModel) {
+                    this.setType(dataModel.type);
+                    this.setFilter(dataModel.filter);
+                }
+                else {
+                    this.setFilter(null);
+                }
+            }
+        };
+    };
+    TextFilterAutoComp.prototype.getApi = function () {
+        return this.api;
+    };
+    return TextFilterAutoComp;
+})();
+
+});
